+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.1.8]
 
 - Update toast for successful swap and add abort controller to fetch assets function.
+- Update the coinbase wallet id.
 
 ## [1.1.7]